mkdir e rmdir pedindo parametros

+++ rmdir.c
@@ -2,13 +2,15 @@
 #include <stdlib.h>
 #include <unistd.h>
 
-int main() {
-    //Variável para armazenar o nome do diretório fornecido pelo usuário
-    char nome_diretorio[100];
+int main(int argc, char *argv[]) {
+    //Verifica se o número de argumentos está correto
+    if (argc != 2) {
+        fprintf(stderr, "Uso: %s <nome_diretorio>\n", argv[0]);
+        return EXIT_FAILURE;
+    }
 
-    //Solicita ao usuário que forneça o nome do diretório a ser criado
-    printf("Digite o nome do diretório a ser apagado: ");
-    scanf("%99s", nome_diretorio);  //Lê o nome do diretório, limitando a 99 caracteres
+    //Obtém o nome do diretório a partir dos argumentos de linha de comando
+    char *nome_diretorio = argv[1];
 
     //Chama a função mkdir() para criar o diretório com o nome fornecido pelo usuário
     int resultado = rmdir(nome_diretorio);
